requirements.txt
Django==1.5.1
Markdown==2.3.1
MySQL-python==1.2.3
PyYAML==3.10
South==0.8.1
amqp==1.0.12
anyjson==0.3.3
billiard==2.7.3.28
boto==2.9.4
celery==3.0.19
distribute==0.6.40
django-celery==3.0.17
django-crispy-forms==1.2.8
django-filter==0.6
django-gravatar==0.1.0
django-jsonfield==0.9.4
django-jsonify==0.2.1
django-model-utils==1.3.1
django-nose==1.1
django-social-auth==0.7.23
djangorestframework==2.3.4
djrill==0.5.0
httplib2==0.8
kombu==2.5.11
lxml==3.2.1
nose==1.3.0
oauth2==1.5.211
python-dateutil==2.1
python-openid==2.2.5
pytz==2013b
requests==1.2.2
six==1.3.0
wsgiref==0.1.2
